# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cek Rekening Bank/E-Wallet API",
        "description": "REST API untuk mengecek nama pemilik rekening bank dan e-wallet Indonesia",
        "contact": {
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login dan dapatkan token API",
                "operationId": "99d66635c4992aeaa6aa44ff653d0563",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login berhasil"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Hapus token API (logout)",
                "operationId": "6abf3b64a4bc7838d56346f05a5153af",
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout berhasil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cek-rekening": {
            "post": {
                "tags": [
                    "Cek Rekening"
                ],
                "summary": "Cek nama pemilik rekening bank/e-wallet",
                "operationId": "d3c0ae5326a9d9071fa3792856ef3477",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bank_code",
                                    "account_number"
                                ],
                                "properties": {
                                    "bank_code": {
                                        "description": "Kode bank/ewallet",
                                        "type": "string",
                                        "example": "bca"
                                    },
                                    "account_number": {
                                        "description": "Nomor rekening",
                                        "type": "string",
                                        "example": "0888123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rekening ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rekening ditemukan"
                                        },
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "valid"
                                                },
                                                "bank_code": {
                                                    "type": "string",
                                                    "example": "bca"
                                                },
                                                "bank_name": {
                                                    "type": "string",
                                                    "example": "Bank BCA"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "bank"
                                                },
                                                "account_number": {
                                                    "type": "string",
                                                    "example": "0888123456"
                                                },
                                                "account_name": {
                                                    "type": "string",
                                                    "example": "Budi Santoso"
                                                },
                                                "reference_id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "validator": {
                                                    "type": "string",
                                                    "example": "dummy"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rekening tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rekening tidak ditemukan"
                                        },
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "invalid"
                                                },
                                                "bank_code": {
                                                    "type": "string",
                                                    "example": "bca"
                                                },
                                                "bank_name": {
                                                    "type": "string",
                                                    "example": "Bank BCA"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "bank"
                                                },
                                                "account_number": {
                                                    "type": "string",
                                                    "example": "xxxx"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Nomor rekening tidak ditemukan"
                                                },
                                                "reference_id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "validator": {
                                                    "type": "string",
                                                    "example": "dummy"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/banks": {
            "get": {
                "tags": [
                    "Cek Rekening"
                ],
                "summary": "Daftar bank dan e-wallet yang didukung",
                "operationId": "bda5feda81006c03b1e60a12e079cba6",
                "responses": {
                    "200": {
                        "description": "Daftar bank berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "bank_code": {
                                                        "type": "string",
                                                        "example": "bca"
                                                    },
                                                    "bank_name": {
                                                        "type": "string",
                                                        "example": "Bank BCA"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "bank"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token API yang didapat dari endpoint login",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Cek Rekening",
            "description": "Cek Rekening"
        }
    ]
}